// frontend/src/api/generated/model/targetKind.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SmartGrao.WebApi | v1
 * OpenAPI spec version: 1.0.0
 */

export type TargetKind = typeof TargetKind[keyof typeof TargetKind];


export const TargetKind = {
  Undefined: 'Undefined',
  Pest: 'Pest',
  FoliarDisease: 'FoliarDisease',
} as const;